fix: delete space under sponsors when there are flags in the screen
