+++ app/page.tsx
@@ -1,4 +1,3 @@
-import SharedSpace from "./shared-space";
 import CursorsContextProvider from "./cursors-context";
 import { Contributors } from "./components/common/Contributors";
 import { Divider } from "./components/common/Divider";
@@ -6,6 +5,8 @@ import { Schedule } from "./components/Schedule";
 import { FAQ } from "./components/FAQ";
 import { Footer } from "./components/Footer";
 import { TicketDownload } from "./components/TicketDownload";
+import Hero from "./components/Hero";
+import Cursors from "./cursors";
 
 export default function Home({
   searchParams,
@@ -16,17 +17,18 @@ export default function Home({
   const room =
     typeof searchParams?.partyroom === "string"
       ? searchParams.partyroom
-      : "voronoi-room";
+      : "aforshow-room";
 
   const host =
     typeof searchParams?.partyhost === "string"
       ? searchParams.partyhost
-      : "voronoi-party.genmon.partykit.dev";
+      : "aforshow-2024-party.jarrisondev.partykit.dev";
 
   return (
-    <main className="font-dmSans">
-      <CursorsContextProvider room={room} host={host}>
-        <SharedSpace />
+    <CursorsContextProvider room={room} host={host}>
+      <main className="font-dmSans">
+        <Cursors />
+        <Hero />
         <section className="flex flex-col gap-[72px]">
           <Divider />
           <Schedule />
@@ -37,7 +39,7 @@ export default function Home({
           <Contributors />
           <Footer />
         </section>
-      </CursorsContextProvider>
-    </main>
+      </main>
+    </CursorsContextProvider>
   );
 }
